feat: add metadata to dashboard, login, profile, settings pages (#353 #354)
Split 4 'use client' pages into thin server page.tsx + XxxClient.tsx:
- dashboard: metadata title/description + robots noindex (private)
- login: metadata + robots noindex
- profile: metadata + robots noindex
- settings: metadata + robots noindex

Each page.tsx is now a Server Component that exports metadata; all hook
usage moves to the co-located *Client.tsx file. Follows the same pattern
as app/page.tsx → HomePageClient.tsx.

Also closes #340 (GoogleAnalytics.tsx already had no 'use client').
